backend/requirements.txt
fastapi
uvicorn
pandas
numpy
feedparser
transformers
torch
tqdm
rich
pytz
matplotlib
seaborn
snscrape
python-dotenv
yfinance
scikit-learn
